Fixes DEVELOPER-3380
Please see https://issues.jboss.org/browse/DEVELOPER-3380 for more
information.

* Fixed composer to use official package location
* Added new twig functions to use in the front end
* Updated styles in topics to prevent side scrolling
* Updated topic layout to send the correct layout for drupal.
