+++ _docker/drupal/drupal-filesystem/composer.json
@@ -9,12 +9,12 @@
             "role": ""
         }
     ],
-    "repositories": [
-        {
+    "repositories": {
+        "drupal": {
             "type": "composer",
-            "url": "https://packagist.drupal-composer.org"
+            "url": "https://packages.drupal.org/8"
         }
-    ],
+    },
     "require": {
         "composer/installers": "1.0.25",
         "drupal-composer/drupal-scaffold": "1.3.2",
@@ -25,7 +25,7 @@
         "drupal/admin_toolbar": "8.1.15",
         "drupal/entity": "8.1.0-alpha3@dev",
         "drupal/entity_storage_migrate": "8.1.x-dev@dev",
-        "drupal/diff": "8.1.0-alpha2",
+        "drupal/diff": "8.1.0-alpha3",
         "drupal/token": "8.1.0-alpha2",
         "drupal/metatag": "8.1.0-beta9",
         "drupal/multiversion": "8.1.0-alpha9",
@@ -39,7 +39,7 @@
         "mikey179/vfsStream": "1.6.3",
         "phpunit/phpunit": "4.8.26",
         "symfony/css-selector": "2.7.13",
-        "drupal/devel": "8.x-1.x-dev@dev"
+        "drupal/devel": "1.x-dev"
     },
     "conflict": {
         "drupal/drupal": "*"
@@ -64,15 +64,25 @@
     },
     "extra": {
         "installer-paths": {
-            "web/core": ["type:drupal-core"],
-            "web/modules/contrib/{$name}": ["type:drupal-module"],
-            "web/profiles/contrib/{$name}": ["type:drupal-profile"],
-            "web/themes/contrib/{$name}": ["type:drupal-theme"],
-            "drush/contrib/{$name}": ["type:drupal-drush"]
+            "web/core": [
+                "type:drupal-core"
+            ],
+            "web/modules/contrib/{$name}": [
+                "type:drupal-module"
+            ],
+            "web/profiles/contrib/{$name}": [
+                "type:drupal-profile"
+            ],
+            "web/themes/contrib/{$name}": [
+                "type:drupal-theme"
+            ],
+            "drush/contrib/{$name}": [
+                "type:drupal-drush"
+            ]
         },
         "patches": {
             "drupal/core": {
-              "Avoid to check the render context to detect early rendering for Non-GET requests.": "2663638-2.patch"
+                "Avoid to check the render context to detect early rendering for Non-GET requests.": "2663638-2.patch"
             }
         }
     }
