+++ _layouts/solution.html.slim
@@ -2,21 +2,22 @@
 layout: technology-base
 ---
 
-- unless site.drupal_base_url
+- if site.drupal_base_url
+  /*.large-20.medium-17.small-24.columns.nav-breakpoint*/
+  = content
+- else 
   section(class="topics-header #{page.solution.id}")
     .row
       .large-24
         h2 #{page.solution.name}
         h4#topic #{page.solution.long_description}
 
-.row.solution-content
-  - unless site.drupal_base_url
+  .row.solution-content
     .large-4.medium-7.small-24.columns.nav-breakpoint
       = partial 'solution-left-menu.html.slim', {'parent' => page}
 
-  .large-20.medium-17.small-24.columns.nav-breakpoint
-    =content
+    .large-20.medium-17.small-24.columns.nav-breakpoint
+      =content
 
-- unless site.drupal_base_url
   = javascripts("solutions", true) do
     script src="#{site.base_url}/javascripts/vendor/chosen.jquery.min.js"
