+++ stylesheets/_topics.scss
@@ -4,7 +4,7 @@
   background-position: top center;
   background-size: cover;
   background-repeat: repeat-x;
-  margin: -15px -15px 0 -15px;
+  margin: -15px 0;
 
   &.containers {background-image: cdn('../images/design/topics/RHDev_topicsheader_desktop_containers.png');}
   &.devops {background-image: cdn('../images/design/topics/RHDev_topicsheader_desktop_devops.png');}
